refactor(batch-sessions): ne plus sauvegarder les instances pour tout le monde
